before changing the copy_sol_to_board funch

+++ IlpSolver.h
@@ -5,7 +5,7 @@
 
 
 int ilp_solver(Game *game) ;
-void copy_sol_to_board(int *sol, Game *game) ;
+void copy_sol_to_board(double *sol, Game *game) ;
 void free_grb(int *ind ,int *indarr2,double *val , double *valarr2, double *sol , char *vtype, GRBenv *env,
               GRBmodel *model,double *lb );
 #endif
